% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklines.R
\name{add_sparkline}
\alias{add_sparkline}
\title{Add a Sparkline to a Column}
\usage{
add_sparkline(lumina_obj, col, color = "#6200ea")
}
\arguments{
\item{lumina_obj}{A lumina widget}

\item{col}{The column name containing numeric data}

\item{color}{The hex color for the bars}
}
\description{
Add a Sparkline to a Column
}
