% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_binding.R
\name{lumina-shiny}
\alias{lumina-shiny}
\alias{luminaOutput}
\alias{renderLumina}
\title{Lumina Table Output for Shiny}
\usage{
luminaOutput(outputId, width = "100\%", height = "auto")

renderLumina(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a lumina table}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Lumina Table Output for Shiny
}
