% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{opt_pagination}
\alias{opt_pagination}
\title{Configure Pagination}
\usage{
opt_pagination(
  w,
  enabled = TRUE,
  pageSize = 10,
  pageSizeOptions = c(10, 25, 50, 100),
  scroller = FALSE,
  showSummary = TRUE
)
}
\arguments{
\item{w}{A lumina widget}

\item{enabled}{Logical. Enable or disable pagination.}

\item{pageSize}{Numeric. The number of rows per page.}

\item{pageSizeOptions}{Numeric vector. The options for the rows per page dropdown.}

\item{scroller}{Logical. Use a scroller instead of pagination?}

\item{showSummary}{Logical. Show or hide the "Showing X to Y of Z results" text.}
}
\description{
Configure Pagination
}
