% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{opt_search}
\alias{opt_search}
\title{Configure Search and Filters}
\usage{
opt_search(
  w,
  show = TRUE,
  highlight = TRUE,
  position = "right",
  filters = FALSE,
  filtersHighlight = TRUE
)
}
\arguments{
\item{w}{A lumina widget}

\item{show}{Logical. Show or hide the global search bar?}

\item{highlight}{Logical. Highlight matched text in search results?}

\item{position}{Character. Position of search bar: "left", "right", "top", or "bottom".}

\item{filters}{Logical. Enable or disable per-column filters?}

\item{filtersHighlight}{Logical. Highlight matched text in column filter results?}
}
\description{
Configure Search and Filters
}
