% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{opt_selection}
\alias{opt_selection}
\title{Configure Row Selection}
\usage{
opt_selection(w, enabled = TRUE, mode = "single", reset = FALSE)
}
\arguments{
\item{w}{A lumina widget}

\item{enabled}{Logical. Enable or disable row selection.}

\item{mode}{Character. Selection mode: "single" or "multiple".}

\item{reset}{Logical. Show a reset button?}
}
\description{
Configure Row Selection
}
