% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklines.R
\name{spark_bar}
\alias{spark_bar}
\title{Generate SVG Sparkline String}
\usage{
spark_bar(x, color = "#6200ea")
}
\arguments{
\item{x}{numeric vector}
}
\description{
Helper function to convert numbers to HTML bars
}
